7.333:SS: permanent integration disable > replace JS alert with waDialog + input

+++ lib/actions/shop/cashShopResetImport.controller.php
@@ -6,14 +6,28 @@
 class cashShopResetImportController extends cashJsonController
 {
     /**
+     * @param null $params
+     *
+     * @return mixed|void
      * @throws waException
-     * @throws Exception
      */
-    public function execute()
+    public function execute($params = null)
     {
         $shopIntegration = new cashShopIntegration();
-        $shopIntegration->getSettings()->resetSettings();
-        $shopIntegration->disableForecast();
-        $shopIntegration->deleteAllShopTransactions();
+
+        $code = wa()->getStorage()->get('cash.shop_integration_reset_code');
+        if (waRequest::getMethod() === 'post') {
+            if ($code === waRequest::post('code', '')) {
+                wa()->getStorage()->del('cash.shop_integration_reset_code');
+
+                $shopIntegration->getSettings()->resetSettings();
+                $shopIntegration->disableForecast();
+                $shopIntegration->deleteAllShopTransactions();
+            } else {
+                $this->setError(_w('Wrong code'));
+            }
+        } else {
+            $this->setError('Wrong method');
+        }
     }
 }
